+++ apps/api/src/console.ts
@@ -1,6 +1,7 @@
 import "reflect-metadata";
 import "@akashnetwork/env-loader";
 import "./open-telemetry";
+import "@src/utils/protobuf";
 
 import { context, trace } from "@opentelemetry/api";
 import { Command } from "commander";
@@ -27,12 +28,24 @@ program
 
 program
   .command("top-up-deployments")
+  .option("-d, --dry-run", "Dry run the top up deployments", false)
   .description("Refill deployments with auto top up enabled")
   .action(async (options, command) => {
     await executeCliHandler(command.name(), async () => {
-      await container.resolve(TopUpDeploymentsController).topUpDeployments();
+      await container.resolve(TopUpDeploymentsController).topUpDeployments({ dryRun: options.dryRun });
     });
   });
+
+program
+  .command("cleanup-stale-deployments")
+  .description("Close deployments without leases created at least 10min ago")
+  .action(async (options, command) => {
+    await executeCliHandler(command.name(), async () => {
+      await container.resolve(TopUpDeploymentsController).cleanUpStaleDeployment();
+    });
+  });
+
+const logger = LoggerService.forContext("CLI");
 
 async function executeCliHandler(name: string, handler: () => Promise<void>) {
   await context.with(trace.setSpan(context.active(), tracer.startSpan(name)), async () => {